+++ index.js
@@ -1,6 +1,6 @@
 const express = require('express')
 const app = express()
-const con = require("./connectionDb")
+const con = require("./config/connectionDb")
 var mysql = require('mysql');
