+++ package.json
@@ -1,17 +1,22 @@
 {
-  "name": "untitled1",
+  "name": "outildevprojet",
   "version": "1.0.0",
-  "main": "index.js",
+  "description": "api rest crud node express mysql",
+  "main": "server.js",
   "scripts": {
-    "start": "node index.js"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "keywords": [],
-  "author": "",
+  "keywords": [
+    "nodejs",
+    "express",
+    "mysql"
+  ],
+  "author": "noeoxycode",
   "license": "ISC",
-  "description": "",
   "dependencies": {
-    "docker": "^1.0.0",
-    "express": "^4.17.2",
+    "cors": "^2.8.5",
+    "express": "^4.17.3",
+    "module": "^1.2.5",
     "mysql": "^2.18.1"
   }
 }
